+++ tests/misc.cpp
@@ -10,6 +10,12 @@ protected:
     std::stringstream ss{};
 };
 
+STEST_F(StreamFixture, TypeInfo) {
+    int i{0};
+    ss << typeid(i);
+    EXPECT_EQ(ss.str(), typeid(i).name());
+}
+
 STEST_F(StreamFixture, PairSameTypes) {
     ss << std::make_pair(0, 1);
     EXPECT_EQ(ss.str(), "{0, 1}");